moved sharing intent back to view
